# src/main/java/com/inigo/fitmentor/coach/infrastructure/api/models/CoachSaveResponse.kt
package com.inigo.fitmentor.coach.infrastructure.api.models

import java.util.Objects
import com.fasterxml.jackson.annotation.JsonProperty
import jakarta.validation.constraints.DecimalMax
import jakarta.validation.constraints.DecimalMin
import jakarta.validation.constraints.Email
import jakarta.validation.constraints.Max
import jakarta.validation.constraints.Min
import jakarta.validation.constraints.NotNull
import jakarta.validation.constraints.Pattern
import jakarta.validation.constraints.Size
import jakarta.validation.Valid
import io.swagger.v3.oas.annotations.media.Schema

/**
 * 
 * @param token Generated
 */
data class CoachSaveResponse(
    @field:NotNull(message = "token must not be null") val token: String? = null
    )