# src/main/resources/openapi/user/coach/creation.yml
openapi: 3.0.1
info:
  title: Auth API
  description: API para gestion de coachs.
  version: 1.0.0
paths:
  /user/coaches:
    put:
      summary: Save coach
      description: Saves coach
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachSaveRequest'
      responses:
        '200':
          description: Coach saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoachSaveResponse'
        '400':
          description: Invalid Request.
        '500':
          description: Server error.
components:
  schemas:
    CoachSaveRequest:
      type: object
      required:
        - user
        - id
      properties:
        user:
          type: UUID
          description: User id
        id:
          type: UUID
          description: Coach id
        name:
          type: string
          description: Name of the coach.
        location:
          type: string
          description: Location of the coach.
        languages:
          type: array
          items:
            type: string
          description: List of languages spoken by the coach.
        physical:
          type: boolean
          description: Offers physical training services.
        experienceYears:
          type: integer
          format: int32
        bio:
          type: string
          description: Coach biography.
        availability:
          type: array
          items:
            $ref: "./common.yml#/TimeSlot"
        clients:
          type: array
          items:
            type: UUID
          description: List of client ids associated with the coach.
        nutritionist:
          type: boolean
          description: Offers nutritionist services.
        trainer:
          type: boolean
          description: Offers trainer services.

    CoachSaveResponse:
      type: object
      properties:
        token:
          type: string
          description: Generated